requirements.txt
Flask
line-bot-sdk

tensorflow
keras
numpy

